requirements.txt
astroid==2.3.3
autopep8==1.4.4
cassandra-driver==3.24.0
certifi==2020.11.8
cffi==1.14.4
click==7.1.2
colorama==0.4.1
cryptography==3.2.1
Flask==1.1.2
geomet==0.2.1.post1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
pyOpenSSL==19.1.0
six==1.15.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.2
